+++ scripts/variables.sh
@@ -1,5 +1,6 @@
 VAR_KEY_PREFIX="@sidebar-key"
 OPTION_DELIMITER="->"
 REGISTERED_PANE_PREFIX="@-sidebar-registered-pane"
+REGISTERED_SIDEBAR_PREFIX="@-sidebar-is-sidebar"
 
 SUPPORTED_TMUX_VERSION="1.9"