+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### master
 - automatic sidebar opening for another sidebar type
+- sidebars can't have nested sidebars
 
 ### v0.1.0, Sep 03, 2014
 - started the project